[CircleCi] Fix tests

+++ geonode/geoserver/manager.py
@@ -163,12 +163,13 @@ class GeoServerResourceManager(ResourceManagerInterface):
                     _src_upload_session = Upload.objects.filter(resource=instance.get_real_instance().resourcebase_ptr)
                     if _src_upload_session.exists():
                         _src_upload_session = _src_upload_session.get()
-                        try:
-                            _src_importer_session = _src_upload_session.get_session.import_session.reload()
-                            importer_session_opts.update({'transforms': _src_importer_session.tasks[0].transforms})
-                        except Exception:
-                            _resource.delete()
-                            raise
+                        if _src_upload_session and _src_upload_session.get_session:
+                            try:
+                                _src_importer_session = _src_upload_session.get_session.import_session.reload()
+                                importer_session_opts.update({'transforms': _src_importer_session.tasks[0].transforms})
+                            except Exception:
+                                _resource.delete()
+                                raise
                 return self.import_dataset(
                     'import_dataset',
                     uuid,
